api/app/requirements.txt
alembic==1.10.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asyncmy==0.2.9
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
elastic-transport==8.15.0
elasticsearch==8.15.1
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.4.2
Mako==1.3.5
MarkupSafe==2.1.5
pika==1.3.2
prompt_toolkit==3.0.47
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.10
pytz==2024.2
redis==5.0.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
